+++ yapyg_movers/control_phys_mover.pxd
@@ -64,4 +64,4 @@ cpdef list create(str entity_name,
                 str neg_sprite,
                 )
 
-cpdef run(list state, str entity_name, list mover, float frame_time_delta, list movers_to_delete)
+cpdef run(list state, str entity_name, list mover, int frame_time_delta, list movers_to_delete)
