+++ yapyg_movers/physical_mover.pyx
@@ -138,7 +138,7 @@ cpdef list create(str entity_name,
                 False
                 ]
 
-cpdef run(list state, str entity_name, list mover, float frame_time_delta, list movers_to_delete):
+cpdef run(list state, str entity_name, list mover, int frame_time_delta, list movers_to_delete):
         """
         TODO
         """
